move ConvosUser into file
